backend/requirements/base.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.7.0
gunicorn==23.0.0
joblib==1.5.2
legacy-cgi==2.6.3
mongoengine==0.29.1
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==3.11.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
celery==5.4.0
redis==5.0.1
